nov11

+++ node2/startercode/adc.h
@@ -2,4 +2,4 @@
 #define ADC_CHANNEL 2
 
 void adc_init();
-uint16_t adc_data();
+uint32_t adc_data();
